fixed the prob of unable to call add and update review api. (ApiService has wrong format, changed to directly use axios)

+++ Server/app/models.py
@@ -32,7 +32,7 @@ class Review(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     text = Column(String)
-    rating = Column(Float)
+    rating = Column(Integer)
     user_id = Column(Integer, ForeignKey("users.id"))
     media_id = Column(Integer, ForeignKey("user_media.id"))
